--- a/baseten-performance-client/Cargo.toml
+++ b/baseten-performance-client/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "baseten_performance_client"
-version = "0.0.5"
+version = "0.0.6"
 edition = "2021"
 
 [dependencies]