--- a/baseten-performance-client/scripts/compare_latency_openai.py
+++ b/baseten-performance-client/scripts/compare_latency_openai.py
@@ -16,7 +16,7 @@ if not api_key:
 api_base_embed = "https://model-yqv4yjjq.api.baseten.co/environments/production/sync"
 
 # Benchmark settings: list of lengths to test.
-benchmark_lengths = [128, 512, 2048, 8192, 32768, 131072]
+benchmark_lengths = [64, 128, 256, 384, 512, 2048, 8192, 32768, 131072]
 micro_batch_size = (
     16  # For AsyncOpenAI client; also used for the PerformanceClient batch
 )
@@ -63,7 +63,7 @@ async def run_baseten_benchmark(length, client=client_b):
 
     # Warm-up run
     _ = await client.async_embed(
-        input=full_input_texts[:512],
+        input=["Hello world"] * 512,
         model="text-embedding-3-small",
         max_concurrent_requests=512,
         batch_size=1,