@michaelfeil 2025-07-09
I'm not sure actually, its just measureing every 0.1s, maybe some kind of measurement that is bad. I think cpu and memory are vastly available. Mostly looking at the times here.